+++ source/engine/makaba/comment/parseAttachments.js
@@ -1,9 +1,9 @@
 import parseAttachment from './parseAttachment'
 
-export default function parseAttachments(post, { toAbsoluteUrl }) {
+export default function parseAttachments(post, { transformAttachmentUrl, toAbsoluteUrl }) {
 	if (post.files.length > 0) {
 		return post.files.map((file) => {
-			return parseAttachment(file, { toAbsoluteUrl })
+			return parseAttachment(file, { transformAttachmentUrl, toAbsoluteUrl })
 		})
 	}
 }